+++ src/createConnectorConfig.ts
